Fixes per eskimo's reviews
